Address review comments
 - refactor templates to use includes
 - use setup() in mixin to initialize once
