+++ caseworker/f680/templates/f680/case/recommendation/includes/recommendation_details_approve.html
@@ -1,0 +1,23 @@
+<div class="govuk-!-padding-4 govuk-!-margin-bottom-2 recommendation-and-decision" data-nosnippet role="region" aria-label="Reason for approving">
+    <div class="govuk-grid-row">
+        <div class="govuk-grid-column-full">
+            <h2 class="govuk-heading-m">Reason for approving</h2>
+            <p class="govuk-body">{{ recommendation.text|linebreaks }}</p>
+
+            {% if recommendation.proviso %}
+            <h2 class="govuk-heading-m">Licence condition</h2>
+            <div class="govuk-body licence-condition">{{ recommendation.proviso|linebreaks }}</div>
+            {% endif %}
+
+            {% if recommendation.note %}
+            <h2 class="govuk-heading-m">Additional instructions</h2>
+            <p class="govuk-body">{{ recommendation.note|linebreaks }}</p>
+            {% endif %}
+
+            {% if recommendation.footnote %}
+            <h2 class="govuk-heading-m">Reporting footnote</h2>
+            <p class="govuk-body">{{ recommendation.footnote|linebreaks }}</p>
+            {% endif %}
+        </div>
+    </div>
+</div>
