+++ caseworker/f680/templates/f680/case/recommendation/includes/recommendation_details_refuse.html
@@ -1,0 +1,13 @@
+<div class="govuk-!-padding-4 govuk-!-margin-bottom-2 recommendation-and-decision" data-nosnippet role="region" aria-label="Reason for refusing">
+    <div class="govuk-grid-row">
+        <div class="govuk-grid-column-full">
+            {% if refusal_note %}
+                <h2 class="govuk-heading-m">Refusal meeting note</h2>
+                <p class="govuk-body">{{ refusal_note.0.text|linebreaks }}</p>
+            {% else %}
+                <h2 class="govuk-heading-m">Reason for refusing</h2>
+                <p class="govuk-body">{{ recommendation.text|linebreaks }}</p>
+            {% endif %}
+        </div>
+    </div>
+</div>
