+++ caseworker/f680/templates/f680/case/recommendation/recommendation_details.html
@@ -16,44 +16,9 @@
 <div class="govuk-grid-row govuk-details">
     <div class="govuk-grid-column-full">
         {% if decision == 'approve' or decision == 'proviso' %}
-        <div class="govuk-!-padding-4 govuk-!-margin-bottom-2 recommendation-and-decision" data-nosnippet role="region" aria-label="Reason for approving">
-            <div class="govuk-grid-row">
-                <div class="govuk-grid-column-full">
-                    <h2 class="govuk-heading-m">Reason for approving</h2>
-                    <p class="govuk-body">{{ recommendation.text|linebreaks }}</p>
-
-                    {% if recommendation.proviso %}
-                    <h2 class="govuk-heading-m">Licence condition</h2>
-                    <div class="govuk-body licence-condition">{{ recommendation.proviso|linebreaks }}</div>
-                    {% endif %}
-
-                    {% if recommendation.note %}
-                    <h2 class="govuk-heading-m">Additional instructions</h2>
-                    <p class="govuk-body">{{ recommendation.note|linebreaks }}</p>
-                    {% endif %}
-
-                    {% if recommendation.footnote %}
-                    <h2 class="govuk-heading-m">Reporting footnote</h2>
-                    <p class="govuk-body">{{ recommendation.footnote|linebreaks }}</p>
-                    {% endif %}
-                </div>
-            </div>
-        </div>
+            {% include "f680/case/recommendation/includes/recommendation_details_approve.html" with recommendation=recommendation %}
         {% elif decision == 'refuse' %}
-        <div class="govuk-!-padding-4 govuk-!-margin-bottom-2 recommendation-and-decision" data-nosnippet role="region" aria-label="Reason for refusing">
-            <div class="govuk-grid-row">
-                <div class="govuk-grid-column-full">
-                {% if refusal_note %}
-                    <h2 class="govuk-heading-m">Refusal meeting note</h2>
-                    <p class="govuk-body">{{ refusal_note.0.text|linebreaks }}</p>
-                {% else %}
-                    <h2 class="govuk-heading-m">Reason for refusing</h2>
-                    <p class="govuk-body">{{ recommendation.text|linebreaks }}</p>
-                {% endif %}
-
-              </div>
-            </div>
-        </div>
+            {% include "f680/case/recommendation/includes/recommendation_details_refuse.html" with recommendation=recommendation %}
         {% endif %}
     </div>
 </div>
